se implementa creacion de LowAmountAccount y las capasidade de  Lista todas las cuentas y Consulta una cuenta por su ID

+++ src/main/java/com/example/patterns_banking/factory/AccountFactory.java
@@ -3,6 +3,8 @@ package com.example.patterns_banking.factory;
 import com.example.patterns_banking.models.Account;
 import com.example.patterns_banking.models.Customer;
 
+import java.math.BigDecimal;
+
 public interface AccountFactory {
-  Account createAccount(Customer customer, String accountNumber, Double balance);
+  Account createAccount(Customer customer, String accountNumber, BigDecimal balance);
 }
