+++ src/main/java/com/example/patterns_banking/services/commands/DepositCommand.java
@@ -3,14 +3,15 @@ package com.example.patterns_banking.services.commands;
 import com.example.patterns_banking.models.Account;
 import com.example.patterns_banking.repositories.IAccountRepository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public class DepositCommand implements ICommand<Account> {
   private final IAccountRepository accountRepository;
   private final Long accountId;
-  private final Double amount;
+  private final BigDecimal amount;
 
-  public DepositCommand(IAccountRepository accountRepository, Long accountId, Double amount) {
+  public DepositCommand(IAccountRepository accountRepository, Long accountId, BigDecimal amount) {
     this.accountRepository = accountRepository;
     this.accountId = accountId;
     this.amount = amount;
